backend/pyproject.toml
[tool.poetry]
name = "scraping-platform-backend"
version = "0.1.0"
description = "Plataforma Paramétrica de Web Scraping Inmobiliario - Backend API"
authors = ["Equipo Desarrollo"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.0"}
asyncpg = "^0.30.0"
alembic = "^1.13.0"
pydantic-settings = "^2.6.0"
celery = {extras = ["redis"], version = "^5.4.0"}
redis = "^5.2.0"
structlog = "^24.1.0"
slowapi = "^0.1.9"
pyjwt = "^2.8.0"
bcrypt = "^4.1.0"
python-multipart = "^0.0.9"
croniter = "^2.0.0"
cron-descriptor = "^1.4.0"
cssselect = "^1.2.0"
playwright = "^1.40.0"
pandas = "^2.2.0"
openpyxl = "^3.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.24.0"
httpx = "^0.27.0"
hypothesis = "^6.100.0"
factory-boy = "^3.3.0"
aiosqlite = "^0.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "property: Property-based tests using Hypothesis",
]

[tool.hypothesis]
# Hypothesis profiles are loaded in tests/conftest.py
